requirements.txt
appdirs==1.4.4
asn1crypto==1.5.1
attrs==25.1.0
backoff==2.2.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
ciso8601==2.3.2
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
filelock==3.17.0
fs==2.4.16
fsspec==2025.2.0
greenlet==3.1.1
idna==3.10
inflection==0.5.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
joblib==1.4.2
jsonpath-ng==1.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==25.6.0
meltanolabs-target-snowflake==0.15.0
more-itertools==10.6.0
packaging==24.2
platformdirs==4.3.6
ply==3.11
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==24.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
rpds-py==0.22.3
setuptools==70.3.0
simpleeval==1.0.3
simplejson==3.19.3
singer-python==6.1.0
singer-sdk==0.44.3
six==1.17.0
snowflake-connector-python==3.13.2
snowflake-sqlalchemy==1.7.3
sortedcontainers==2.4.0
SQLAlchemy==2.0.38
tomlkit==0.13.2
typing_extensions==4.12.2
urllib3==2.3.0
